+++ tests/test_exceptions.py
@@ -634,3 +634,63 @@ class TestSGLangExceptionMap:
         import openai
         result = normalize_provider_exception(Mock(spec=openai.RateLimitError), "sglang")
         assert result.provider == "sglang"
+
+
+# ---------------------------------------------------------------------------
+# Local model error passthrough (Transformers, MLXLM, VLLMOffline)
+# ---------------------------------------------------------------------------
+
+class TestLocalModelErrorPassthrough:
+    def test_transformers_passthrough_native_generation_errors(self):
+        from outlines.models.transformers import Transformers
+
+        class DummyModel:
+            class config:
+                is_encoder_decoder = False
+
+            def generate(self, **kwargs):
+                raise RuntimeError("local generation failed")
+
+        model = Transformers.__new__(Transformers)
+        model.model = DummyModel()
+
+        with pytest.raises(RuntimeError, match="local generation failed"):
+            model._generate_output_seq("", {"input_ids": object()})
+
+    def test_mlxlm_passthrough_native_generation_errors(self, monkeypatch):
+        import sys
+        import types
+        from outlines.models.mlxlm import MLXLM
+
+        def _raise_generate(*args, **kwargs):
+            raise ValueError("mlx local failure")
+
+        fake_mlx_lm = types.SimpleNamespace(generate=_raise_generate)
+        monkeypatch.setitem(sys.modules, "mlx_lm", fake_mlx_lm)
+
+        model = MLXLM.__new__(MLXLM)
+        model.model = object()
+        model.mlx_tokenizer = object()
+        model.type_adapter = types.SimpleNamespace(
+            format_input=lambda _: "prompt",
+            format_output_type=lambda _: None,
+        )
+
+        with pytest.raises(ValueError, match="mlx local failure"):
+            model.generate("hello")
+
+    def test_vllm_offline_passthrough_native_generation_errors(self):
+        import types
+        from outlines.models.vllm_offline import VLLMOffline
+
+        class DummyModel:
+            def generate(self, **kwargs):
+                raise RuntimeError("vllm offline failed")
+
+        model = VLLMOffline.__new__(VLLMOffline)
+        model.model = DummyModel()
+        model.type_adapter = types.SimpleNamespace(format_input=lambda _: "prompt")
+        model._build_generation_args = lambda _inference_kwargs, _output_type: object()
+
+        with pytest.raises(RuntimeError, match="vllm offline failed"):
+            model.generate("hello")
